feat：Move `URLProtocolStub` to separate file
